+++ src/lib.rs
@@ -1,6 +1,3 @@
 pub mod btc;
 pub mod mempool;
 pub mod util;
-
-#[cfg(test)]
-mod test;
